update test case

+++ shardingsphere-scenario/src/main/java/test/apache/skywalking/testcase/shardingsphere/service/utility/algorithm/PreciseModuloShardingTableAlgorithm.java
@@ -23,10 +23,10 @@ import io.shardingsphere.api.algorithm.sharding.standard.PreciseShardingAlgorith
 
 import java.util.Collection;
 
-public class PreciseModuloShardingTableAlgorithm implements PreciseShardingAlgorithm<Long> {
+public class PreciseModuloShardingTableAlgorithm implements PreciseShardingAlgorithm<Integer> {
     
     @Override
-    public String doSharding(final Collection<String> tableNames, final PreciseShardingValue<Long> shardingValue) {
+    public String doSharding(final Collection<String> tableNames, final PreciseShardingValue<Integer> shardingValue) {
         for (String each : tableNames) {
             if (each.endsWith(shardingValue.getValue() % 2 + "")) {
                 return each;
